requirements.txt
# WEIS - Wollam Estimating Intelligence System
# Phase 1 Dependencies

# AI / LLM
anthropic>=0.39.0

# Environment
python-dotenv>=1.0.0

# Database (SQLite is stdlib, but we use this for migrations)
# No additional DB package needed for SQLite

# CLI Interface
rich>=13.7.0

# Web Interface
streamlit>=1.28.0

# Data Processing
# (stdlib re, csv, json are sufficient for JCD parsing)

# Document Processing (Phase 2.4 — Active Bid Documents)
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0

# HCSS API Integration (Phase A)
httpx>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0

# NiceGUI Web Interface
nicegui>=3.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
